Move tree/tree.py implementation into ui/wx
