+++ pyface/tree/api.py
@@ -1,5 +1,5 @@
 #------------------------------------------------------------------------------
-# Copyright (c) 2005-2011, Enthought, Inc.
+# Copyright (c) 2005-2017, Enthought, Inc.
 # All rights reserved.
 #
 # This software is provided without warranty under the terms of the BSD
@@ -24,9 +24,5 @@ from .trait_dict_node_type import TraitDictNodeType
 from .trait_list_node_type import TraitListNodeType
 from .tree_model import TreeModel
 
-from traits.etsconfig.api import ETSConfig
-if ETSConfig.toolkit == 'wx':
-    # Tree has not yet been ported to qt
-    from .tree import Tree
-
-del ETSConfig
+# Tree has not yet been ported to qt
+from .tree import Tree
